requirements.txt
chromadb>=0.5.0
pymupdf>=1.24.0
python-dotenv>=1.0.0
llama-cpp-python>=0.2.75
sentence-transformers>=2.7.0
huggingface_hub>=0.23.0
